chore: move private methods from default info export
