+++ release-please-config.json
@@ -8,7 +8,7 @@
       "draft": false,
       "prerelease": false,
       "include-component-in-tag": false,
-      "extra-files": ["src/lib/core/info.ts"]
+      "extra-files": ["src/lib/core/info/version.ts"]
     }
   },
   "$schema": "https://raw.githubusercontent.com/googleapis/release-please/main/schemas/config.json"
