+++ src/lib/core/info/version.ts
@@ -1,0 +1,3 @@
+// x-release-please-start-version
+export const VERSION = '3.3.1';
+// x-release-please-end
